fix: requirement에 isAllverified 옮기기
